requirements.txt
djangorestframework==3.16.1
Django==5.2.5
django-filter==25.1